[bug fix] repeat count info is not working.
The repleatCount in png metadata means that repeat animation after end of the last play.
So #0 means that 1 time animation from first frame.
#1 means that 2 times repeat.
#2 means that 3 time repeat.

REF. https://wiki.mozilla.org/APNG_Specification
`num_plays` indicates the number of times that this animation should play; if it is 0, the animation should play indefinitely. If nonzero, the animation should come to rest on the final frame at the end of the last play.